make region optional
